chore: extract demo-mode fixtures out of pc.service
Move DEMO_PCS seed data and the pure demo-only level-up shim (slot/feat/
spell/subclass tables and their helper functions) into a sibling
pc.demo-data.ts. pc.service.ts now reads as CRUD/HTTP logic only; the two
stateful demo orchestrators (computeDemoPreview, applyDemoLevelUp) stay in
the service and delegate to the imported pure functions.

Pure relocation, no behavior change: ng build green, full unit suite
unchanged (239 pass; the 4 pre-existing CreateCharacterModalComponent
failures are unrelated). Service drops 707 -> ~430 lines.
